@@ -22,10 +22,10 @@
         "silverstripe/sentry": "*"
     },
     "require": {
-        "php": "^7||^8",
+        "php": "^8.3",
         "sentry/sdk": "^3",
         "composer/package-versions-deprecated": "^1.11",
-        "silverstripe/framework": "^5"
+        "silverstripe/framework": "^6"
     },
     "require-dev": {
         "phpunit/phpunit": "^9"
